chore: reorder functions and norm
