+++ Makefile
@@ -6,7 +6,7 @@
 #    By: pmarkaid <pmarkaid@student.hive.fi>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/10/28 11:57:01 by dbejar-s          #+#    #+#              #
-#    Updated: 2024/11/21 16:27:42 by pmarkaid         ###   ########.fr        #
+#    Updated: 2024/11/22 13:56:15 by pmarkaid         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -20,6 +20,7 @@ SRCS = \
 	input.c \
 	input_utils.c \
 	input_parse.c \
+	input_parse_fc.c \
 	validation.c \
 	validation_map.c \
 	hooks.c \
